refactor: Move merge logic into timetable.cc
The merge logic doesn't have much to do with loading timetables in
general and is heavily dependent on the structure of timetables.
Thus, move the merge logic into a function inside the timetable class.

There is no change in behaviour expected. For completeness, a description
of the structure of the new method is provided:

1. Some fields don't get used during loading and always remain empty. They
don't need to be merged and aren't handled. This precondition is checked in
case this assumption changes in the future.
2. date_range must be the same between the timetables. It is set from the
outside and it is assumed that at least one timetable has the correct
date_range (and then all of them have if no errors occur).
3. This isn't strictly necessary, but does provide a quick overview of
which fields are handled. They are all read-only of course. The order is
based on the order of definition in timetable.h
4. bitfields and string_idx map have to be merged first to determine the
new indices. They are then passed to the index mapping.
5. The index mapping is a read-only structure which determines the mapping
from the (relative) indices in other_tt to the new absolute indexing.
6. The other fields get merged independently from each other. The mapping
indices only depend on some offsets that can be statically determined
beforehand.

The merging is done structurally, with some exceptions. Each field gets merged
according to its type, until objects are reached which don't need merging as
they stand alone. However, they might contain or be references which need
redirecting to new targets since the merge invalidates most references within
other_tt. This too can be done mostly structurally by mapping the members of
composite types and mapping affected reference types. Care must be taking with
possibly invalid references as they must remain invalid after mapping.
